requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
torchvision==0.15.2+cpu
alembic==1.13.2
boto3==1.34.145
botocore==1.34.145
Flask==3.0.3
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
Jinja2==3.1.4
Mako==1.1.4
MarkupSafe==2.1.5
networkx==3.3
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.10.0.84
protobuf==5.27.2
psycopg2-binary==2.9.1
python-dotenv==1.0.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
